+++ src/test/unit/actions.test.ts
@@ -1,7 +1,7 @@
 import { resolve, join } from 'path';
-import * as assert from 'assert';
-import * as sinon from 'sinon';
-import * as proxyquire from 'proxyquire';
+import assert from 'assert';
+import sinon from 'sinon';
+import proxyquire from 'proxyquire';
 
 const sandbox = sinon.createSandbox();