+++ src/test/unit/helpers/error-message.test.ts
@@ -1,6 +1,6 @@
-import * as assert from 'assert';
-import * as sinon from 'sinon';
-import * as proxyquire from 'proxyquire';
+import assert from 'assert';
+import sinon from 'sinon';
+import proxyquire from 'proxyquire';
 
 describe('Error Helper', () => {
   let errorMessage;